AlgoBulls_dev_proj/ToDo/requirements.txt
anyio==3.6.2
asgiref==3.7.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
Django==5.0
djangorestframework==3.14.0
dnspython==2.3.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
networkx==3.2.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==1.10.7
pymongo==4.3.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-mock==1.10.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
SQLAlchemy==1.4.47
sqlparse==0.4.4
starlette==0.26.1
threadpoolctl==3.1.0
tweepy==4.13.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.3
